+++ optimesh/cvt/helpers.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+#
+
+
+def jac_uniform(mesh):
+    # create Jacobian
+    centroids = mesh.control_volume_centroids
+    X = mesh.node_coords
+    jac = 2 * ((X - centroids).T * mesh.control_volumes).T
+    return jac.flatten()
